Starting to split core.py into more manageable modules.
