+++ pywr/domains/river.py
@@ -1,5 +1,5 @@
 
-from ..core import Node, Domain, Input, Output, Link, Storage, PiecewiseLink, MultiSplitLink
+from pywr.nodes import Node, Domain, Input, Output, Link, Storage, PiecewiseLink, MultiSplitLink
 from pywr.parameters import pop_kwarg_parameter, ConstantParameter, BaseParameter, load_parameter
 from pywr.parameters.control_curves import ControlCurveParameter
